Adds additional attribute collector properties. Remove attr-appcompat in favor of doing support lib views within attr plugin itself.

+++ hyperion-core/src/main/java/com/willowtreeapps/hyperion/core/AttributeTranslator.java
@@ -8,6 +8,6 @@ public interface AttributeTranslator {
 
     String translatePx(@Px int px);
 
-    String translatePxToSp(float px);
+    String translatePxToSp(int px);
 
 }
